Pfade in tool scripts relativ gesetzt

+++ tools/backup_db.sh
@@ -1,12 +1,33 @@
 all_tables=( "TBL_MOVIE" "TBL_ACTOR" "TBL_MOVIE_ACTOR" "TBL_LANGUAGE" "TBL_MOVIE_LANGUAGE" "TBL_MOVIE_SUBTITLE" "TBL_SONG" "TBL_MOVIE_SONG" "TBL_WATCH" "TBL_GENRE" "TBL_MOVIE_GENRE" )
 dir_name=$(date +"%d_%m_%Y_T_%H_%M_%S")
 
-pushd /home/wolfgang/Projekte/Movie/database/backup
+user=$(whoami)
+
+if [ "$user" != "root" ]
+then
+  echo "This script has to be run as root to make use of -my.cnf!"
+  exit 1
+fi
+
+# current directory of caller
+current_dir=$(pwd)
+
+# directory given in the call
+call_path=$(dirname "$0")
+
+# absolute path to the tools directory containing this script
+tools_path=$current_dir/$call_path
+
+pushd $tools_path/../database/backup >/dev/null
 
 
-# delete oldest backups, so that there is a maximum of 9 old backups left
-echo "deleting old backups ..."
-ls -t | sed -e '1,9d' | xargs -d '\n' rm -r
+backup_count=$(ls -t | wc -l)
+if [ $backup_count -gt 9 ]
+then
+  # delete oldest backups, so that there is a maximum of 9 old backups left
+  echo "deleting old backups ..."
+  ls -t | sed -e '1,9d' | xargs -d '\n' rm -r
+fi
 
 # create current backup directory, this is the 10th backup
 mkdir $dir_name
@@ -18,15 +39,15 @@ then
   for table in "$@"
   do
     echo "dumping specific table: $table"
-    mysqldump -u movie --password=P4l.xA#3W?s_ --no-create-db --no-create-info --compact --complete-insert --quick --disable-keys MovieDB_1_0 $table | gzip > $dir_name/dump_$table.sql.gz
+    mysqldump --no-create-db --no-create-info --compact --complete-insert --quick --disable-keys --databases MovieDB_1_0 --tables $table | gzip > $dir_name/dump_$table.sql.gz
   done
 else
   echo "dumping all tables:"
   for table in "${all_tables[@]}"
   do
     echo "  $table"
-    mysqldump -u movie --password=P4l.xA#3W?s_ --no-create-db --no-create-info --compact --complete-insert --quick --disable-keys MovieDB_1_0 $table | gzip > $dir_name/dump_$table.sql.gz
+    mysqldump --no-create-db --no-create-info --compact --complete-insert --quick --disable-keys --databases MovieDB_1_0 --tables $table | gzip > $dir_name/dump_$table.sql.gz
   done
 fi
 
-popd
+popd >/dev/null
